# lib/api-zod/src/generated/types/adminProduct.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminProduct {
  id: number;
  name: string;
  brand: string;
  categoryId: string;
  /** @nullable */
  subcategory?: string | null;
  /** @nullable */
  description?: string | null;
  mrp: number;
  price: number;
  weight: string;
  /** @nullable */
  imageUrl?: string | null;
  stockQty: number;
  inStock: boolean;
  enabled: boolean;
  isBestSeller: boolean;
  isDwarikaSpecial: boolean;
  createdAt: string;
}