# lib/api-zod/src/generated/types/banner.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Banner {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  buttonText?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  bg?: string | null;
  enabled: boolean;
  sortOrder: number;
  createdAt: string;
}