# lib/api-zod/src/generated/types/bannerUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BannerUpdate {
  title?: string;
  subtitle?: string;
  buttonText?: string;
  imageUrl?: string;
  bg?: string;
  enabled?: boolean;
  sortOrder?: number;
}