# lib/api-zod/src/generated/types/bulkImportInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInput } from './productInput';

export interface BulkImportInput {
  products: ProductInput[];
}