# lib/api-zod/src/generated/types/bulkImportResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BulkImportResult {
  imported: number;
  updated: number;
  failed: number;
  errors: string[];
}