# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderAddress } from './orderAddress';
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  orderNumber: string;
  /** @nullable */
  customerId?: number | null;
  customerName: string;
  customerMobile: string;
  address: OrderAddress;
  items: OrderItem[];
  subtotal: number;
  smallCartFee: number;
  deliveryFee: number;
  handlingFee: number;
  grandTotal: number;
  paymentMethod: string;
  /** @nullable */
  orderNote?: string | null;
  status: string;
  createdAt: string;
}