# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  /** @nullable */
  productId?: number | null;
  name: string;
  brand: string;
  weight: string;
  qty: number;
  price: number;
}